pyproject.toml
[project]
name = "project1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "huggingface-hub>=0.33.1",
    "langchain>=0.3.26",
    "langchain-anthropic>=0.3.16",
    "langchain-community>=0.3.26",
    "langchain-google-genai>=2.1.5",
    "langchain-huggingface>=0.3.0",
    "langchain-openai>=0.3.26",
    "langgraph>=0.5.0",
    "python-dotenv>=1.1.1",
]
